Allow chart metrics to be configured per dataset analysed
Fix #669
